Document how to setup Union Scope.

+++ visual_translations/management/commands/generate_eu_visual_translations.py
@@ -7,11 +7,17 @@ class Command(GrowCommunityCommand):
 
     community_model = "VisualTranslationsEUCommunity"
 
+    def add_arguments(self, parser):
+        parser.add_argument('community', type=str, nargs="?", default=self.community_model)
+        parser.add_argument('-a', '--args', type=str, nargs="*", default="")
+        parser.add_argument('-d', '--delete', action="store_true")
+
     def get_community(self):
         return self.model.objects.create_by_signature(self.signature, **self.config)
 
-    def handle(self, *args, **kwargs):
-        GoogleTranslate.objects.all().delete()
-        GoogleImage.objects.all().delete()
-        super(Command, self).handle(*args, **kwargs)
+    def handle(self, *args, **options):
+        if options["delete"]:
+            GoogleTranslate.objects.all().delete()
+            GoogleImage.objects.all().delete()
+        super(Command, self).handle(*args, **options)
         self.model.objects.delete_manifestations_by_signature(signature=self.signature)
